requirements.txt
asgiref==3.7.1
boto3==1.26.140
botocore==1.29.140
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-storages==1.13.2
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
mysqlclient==2.1.1
packaging==23.1
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
